chore: move service mappers to structures file
